REFACTOR UiPageInstaller UIPageError Namespace angepasst
