+++ Exceptions/UiPage/UiPageCreateError.php
@@ -1,5 +1,5 @@
 <?php
-namespace exface\Core\Exceptions;
+namespace exface\Core\Exceptions\UiPage;
 
 /**
  * Error thrown when a problem creating a UiPage occurs.
