+++ Exceptions/UiPage/UiPageIdNotUniqueError.php
@@ -1,5 +1,5 @@
 <?php
-namespace exface\Core\Exceptions;
+namespace exface\Core\Exceptions\UiPage;
 
 /**
  * Exception thrown if the ID of a UI page is not unique.
